refactor(api): move db integration tests in its own folder
