+++ api/tests/db/integration/infrastructure/database-connection_test.js
@@ -1,12 +1,12 @@
-import datamartKnexConfigs from '../../../datamart/knexfile.js';
-import datawarehouseKnexConfigs from '../../../datawarehouse/knexfile.js';
-import { DatabaseConnection } from '../../../db/database-connection.js';
-import liveKnexConfigs from '../../../db/knexfile.js';
-import * as userRepository from '../../../src/identity-access-management/infrastructure/repositories/user.repository.js';
-import { config } from '../../../src/shared/config.js';
-import { UserNotFoundError } from '../../../src/shared/domain/errors.js';
-import { expect } from '../../test-helper.js';
-import { databaseBuilder } from '../../tooling/databases.js';
+import datamartKnexConfigs from '../../../../datamart/knexfile.js';
+import datawarehouseKnexConfigs from '../../../../datawarehouse/knexfile.js';
+import { DatabaseConnection } from '../../../../db/database-connection.js';
+import liveKnexConfigs from '../../../../db/knexfile.js';
+import * as userRepository from '../../../../src/identity-access-management/infrastructure/repositories/user.repository.js';
+import { config } from '../../../../src/shared/config.js';
+import { UserNotFoundError } from '../../../../src/shared/domain/errors.js';
+import { expect } from '../../../test-helper.js';
+import { databaseBuilder } from '../../../tooling/databases.js';
 
 describe('Integration | Infrastructure | database-connection', function () {
   describe('#emptyAllTables', function () {
